requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
assemblyai
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
croniter==6.0.0
cryptography==45.0.6
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.12.0
firebase_admin==7.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
fonttools==4.59.1
fpdf==1.7.2
fpdf2==2.8.4
frozenlist==1.7.0
git-filter-repo==2.47.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==22.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
icalendar==6.3.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langsmith==0.4.16
lxml==6.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
moviepy==2.2.1
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.2.3
openai==1.100.2
orjson==3.11.2
packaging==25.0
pillow==11.3.0
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
psutil==6.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.5.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stripe==12.5.0
tenacity==9.1.2
twilio==9.3.5
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zstandard==0.24.0
